Catch LogSender exception and comment out some prints

+++ Hermes/src/main/java/analyzer/extension/replayView/DistillerReplayer.java
@@ -16,6 +16,7 @@ import ch.uzh.ifi.seal.changedistiller.ChangeDistiller;
 import ch.uzh.ifi.seal.changedistiller.ChangeDistiller.Language;
 import ch.uzh.ifi.seal.changedistiller.distilling.FileDistiller;
 import ch.uzh.ifi.seal.changedistiller.model.entities.SourceCodeChange;
+import ch.uzh.ifi.seal.changedistiller.model.entities.SourceCodeEntity;
 import fluorite.commands.CompilationEventCommand;
 import fluorite.commands.Delete;
 import fluorite.commands.EHICommand;
@@ -37,11 +38,14 @@ public class DistillerReplayer extends ADifficultyPredictionAndStatusPrinter{
 	private String fileContent;
 	static final long DAY = 24 * 3600 * 1000; 
 	static FileDistiller distiller = ChangeDistiller.createFileDistiller(Language.JAVA);
+	static File f1 = new File("E:\\Test\\524A5\\replaySrc0\\main\\lisp\\BasicOperationManager.java");
+	static File f2 = new File("E:\\Test\\524A5\\replaySrc0\\main\\lisp\\evaluator\\BasicOperationManager.java");
 
 	public static void main(String[] args) {
 //		printChangesBetweenCompile();
 		printChangesBetweenSave();
 //		printChangesByTime();
+//		System.out.println(findFileChanges(f1, f2));
 	}
 	
 	public static void printChangesByTime() {
@@ -99,16 +103,53 @@ public class DistillerReplayer extends ADifficultyPredictionAndStatusPrinter{
 		for (Map<String, List<SourceCodeChange>> changes: replayer.findChangesBetweenSave(PROJECT, 0, Long.MAX_VALUE)) {
 			if (changes == null) return;
 			System.out.println("Save " + count++ + ":" + System.lineSeparator());
+//			System.out.println(changes);
 			for (String key : changes.keySet()) {
 				if (changes.get(key).isEmpty()) {
 					continue;
 				}
 				System.out.println(key);
 				for (SourceCodeChange change : changes.get(key)) {
-					System.out.println(change.getClass().getSimpleName() + "change:" + change.getChangedEntity() + " " + change.getChangeType());
+					System.out.println(change.getChangeType().toString() + ": " + change.getChangedEntity() + System.lineSeparator() + "\tParent = " + change.getParentEntity() + " range = " + change.getChangedEntity().getSourceRange());
+//					System.out.println("**********************************************");
+//					System.out.println("toString() = " + change.toString());
+//					System.out.println("getClass() = " + change.getClass().getSimpleName());
+//					System.out.println("getSignificanceLevel() = " + change.getSignificanceLevel());
+//					System.out.println("getLabel() = " + change.getLabel());
+//					SourceCodeEntity entity = change.getChangedEntity();
+//					System.out.println();
+//					System.out.println("changedEntity = " + entity);
+//					System.out.println("entity.getLabel() = " + entity.getLabel());
+//					System.out.println("entity.getModifiers() = " + entity.getModifiers());
+//					System.out.println("entity.getUniqueName() = " + entity.getUniqueName());
+//					System.out.println("entity.getAssociatedEntities() = " + entity.getAssociatedEntities());
+//					System.out.println("entity.getSourceRange() = " + entity.getSourceRange());
+//					System.out.println("entity.getType() = " + entity.getType());
+//					
+//					ChangeType changeType = change.getChangeType();
+//					System.out.println();
+//					System.out.println("changedType = " + changeType);
+//					entity = change.getParentEntity();
+//					System.out.println();
+//					System.out.println("parentEntity = " + entity);
+//					System.out.println("parentEntity.getLabel() = " + entity.getLabel());
+//					System.out.println("parentEntity.getModifiers() = " + entity.getModifiers());
+//					System.out.println("parentEntity.getUniqueName() = " + entity.getUniqueName());
+//					System.out.println("parentEntity.getAssociatedEntities() = " + entity.getAssociatedEntities());
+//					System.out.println("parentEntity.getSourceRange() = " + entity.getSourceRange());
+//					System.out.println("parentEntity.getType() = " + entity.getType());
+//					StructureEntityVersion rootEntity = change.getRootEntity();
+//					System.out.println("rootEntity = " + rootEntity);
+//					System.out.println("rootEntity.getUniqueName() = " + rootEntity.getUniqueName());
+//					System.out.println("rootEntity.getVersion() = " + rootEntity.getVersion());
+//					System.out.println("rootEntity.getSourceCodeChanges() = " + rootEntity.getSourceCodeChanges());
+//					System.out.println("rootEntity.getType() = " + rootEntity.getType());
+//					System.out.println();
+//					
 //					System.out.println(change.toString());
 				}
 				System.out.println();
+//				return;
 			}
 		}
 		System.exit(0);
@@ -117,8 +158,9 @@ public class DistillerReplayer extends ADifficultyPredictionAndStatusPrinter{
 	public List<Map<String, List<SourceCodeChange>>> findChangesBetweenSave(String project, long startTime, long endTime) {
 		File[] src = initProject(PROJECT);
 		replayTo(src[currentSrc], startTime);
-		List<Map<String, List<SourceCodeChange>>> changesBetweenCompile = new ArrayList<>();
+		List<Map<String, List<SourceCodeChange>>> changesBetweenSave = new ArrayList<>();
 		long currentTime = 0;
+		boolean empty = false;
 		for (int i = 0; i < nestedCommands.size(); i++) {
 			List<EHICommand> commands = nestedCommands.get(i);
 			for (int j = 0; j < commands.size(); j++) {
@@ -128,20 +170,42 @@ public class DistillerReplayer extends ADifficultyPredictionAndStatusPrinter{
 						currentTime = command.getStartTimestamp()+command.getTimestamp();
 						File src1 = src[currentSrc % 2];
 						File src2 = src[(currentSrc + 1) % 2];
-						if (src2.exists()) {
-							FileUtility.deleteFolder(src2);
+						if (!empty) {
+							if (src2.exists()) {
+								FileUtility.deleteFolder(src2);
+							}
+							FileUtility.copyDirectory(src1.getPath(), src2.getPath());
 						}
-						FileUtility.copyDirectory(src1.getPath(), src2.getPath());
-						replayTo(src[(currentSrc + 1) % 2], currentTime);
-						changesBetweenCompile.add(findProjectChanges(src1, src2));
+//						if (src2.exists()) {
+//							FileUtility.deleteFolder(src2);
+//						}
+//						FileUtility.copyDirectory(src1.getPath(), src2.getPath());
+						replayTo(src2, currentTime);
+						Map<String, List<SourceCodeChange>> changes = findProjectChanges(src1, src2);
+						empty = true;
+						for (List<SourceCodeChange> change : changes.values()) {
+							if (!change.isEmpty()) {
+								empty = false;
+								break;
+							}
+						}
+						if (!empty) {
+//							System.out.println("is empty");
+							changesBetweenSave.add(changes);
+							currentSrc++;
+						}
+//						changesBetweenCompile.add(findProjectChanges(src1, src2));
 					} catch (IOException e) {
 						e.printStackTrace();
 					}
-					currentSrc++;
+//					currentSrc++;
 				}
 			}
 		}
-		return changesBetweenCompile;
+		for (File srcFolder: src) {
+			FileUtility.deleteFolder(srcFolder);
+		}
+		return changesBetweenSave;
 	}
 	
 	public List<Map<String, List<SourceCodeChange>>> findChangesBetweenCompile(String project, long startTime, long endTime) {
@@ -282,11 +346,57 @@ public class DistillerReplayer extends ADifficultyPredictionAndStatusPrinter{
 		}
 		return allChanges;
 	}
-
-	public List<SourceCodeChange> findFileChanges(File file1, File file2) {
+	
+	static final String ADD_FUN = "ADDITIONAL_FUNCTIONALITY";
+	
+	public static List<SourceCodeChange> findFileChanges(File file1, File file2) {
 		try {
 			distiller.extractClassifiedSourceCodeChanges(file1, file2);
 			List<SourceCodeChange> changes = distiller.getSourceCodeChanges();
+			List<String> newMethods = new ArrayList<>();
+			StringBuilder content = FileUtility.readFile(file2);
+			for (SourceCodeChange change: changes) {
+				if (change.getLabel().equals(ADD_FUN)) {
+					SourceCodeEntity entity = change.getChangedEntity();
+					String method = content.substring(entity.getStartPosition(), entity.getEndPosition());
+					if (entity.isAbstract()) {
+						method += ";";
+					} else {
+						method = method.substring(0, method.indexOf("{"));
+						newMethods.add(method+"{}");
+					}
+				}
+			}
+			if (!newMethods.isEmpty()) {
+				content = FileUtility.readFile(file1);
+				content.deleteCharAt(content.lastIndexOf("}"));
+				for (String method: newMethods) {
+					content.append(System.lineSeparator() + method);
+				}
+				content.append(System.lineSeparator() + "}");
+				try (BufferedWriter bw = new BufferedWriter(new FileWriter(file1))) {
+					bw.write(content.toString());
+				}
+				distiller.extractClassifiedSourceCodeChanges(file1, file2);
+				List<SourceCodeChange> methodChanges = distiller.getSourceCodeChanges();
+				if (!methodChanges.isEmpty()) {
+					for (int i = 0, j = 0; i < changes.size(); i++) {
+						SourceCodeChange change = changes.get(i);
+						if (change.getLabel().equals(ADD_FUN)) {
+							while (j < methodChanges.size() && !methodChanges.get(j).getRootEntity().getUniqueName().equals(change.getChangedEntity().getUniqueName())) {
+								j++;
+							}
+							if (methodChanges.size() < j) {
+								methodChanges.add(j, change);
+								j++;
+							} else {
+								methodChanges.add(change);
+							}
+						}
+					}
+					return methodChanges;
+				}
+			}
 			return changes;
 		} catch (Exception e) {
 			System.err.println("Warning: error while change distilling. " + e.getMessage());
@@ -294,9 +404,9 @@ public class DistillerReplayer extends ADifficultyPredictionAndStatusPrinter{
 		}
 		return null;
 	}
-
-	public void replayTo(File src, long endTime) {
+	
 		FileEditor editor;
+		public void replayTo(File src, long endTime) {
 		if (fileContent == null) {
 			editor = FileEditor.getEditor(src, currentFileName);
 		} else {
